move values to a separate file
